feature/bo-16 фиксы по конфликтам
