+++ packages/client/package.json
@@ -3,7 +3,7 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "dev": "vite",
+    "dev": "vite --force",
     "build": "tsc && vite build",
     "preview": "vite preview",
     "lint": "eslint .",
@@ -11,12 +11,17 @@
     "test": "jest ./"
   },
   "dependencies": {
+    "@emotion/react": "11.10.5",
+    "@emotion/styled": "11.10.5",
+    "@mui/material": "5.11.5",
     "dotenv": "^16.0.2",
     "eslint-config-prettier": "^8.5.0",
+    "formik": "2.2.9",
     "prettier": "^2.7.1",
     "react": "^18.2.0",
     "react-dom": "18.2.0",
-    "react-router-dom": "6.6.1"
+    "react-router-dom": "6.6.1",
+    "yup": "0.32.11"
   },
   "devDependencies": {
     "@testing-library/react": "^13.3.0",
