--- packages/client/src/Pages/Signup/Signup.module.css
@@ -1,20 +1,0 @@
-.mainWrapper {
-  height: 100vh;
-  display: flex;
-}
-.mainBackground {
-  width: 50%;
-  height: 100%;
-  background-image: url("./img/bomberman.webp");
-  background-size: cover;
-  background-position: center;
-  background-repeat: no-repeat;
-}
-.loginWrapper {
-  width: 50%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 35px;
-  justify-content: center;
-}
